Изменение типа цены товара на Long
В будущем тип цены товара будет изменен на BigDecimal

+++ src/main/java/ru/geekbrains/webstore/dtos/OrderDto.java
@@ -14,7 +14,7 @@ public class OrderDto {
 
   @NotNull(message = "Order must have a purchase price")
   @Min(value = 1, message = "Purchase price must be greater than 0")
-  private Double purchasePrise;
+  private Long purchasePrise;
 
   @NotNull(message = "User must have a name")
   private String userName;
